chore(interactive notifications): Move to enum
